+++ MovieQuiz/Services/MovieQuizViewControllerProtocol.swift
@@ -11,4 +11,5 @@ protocol MovieQuizViewControllerProtocol: AnyObject {
     func hideLoadingIndicator()
     
     func showNetworkError(message: String)
+    func showAlert(model: AlertModel)
 }
