+++ MovieQuiz/Services/MoviesLoader.swift
@@ -6,6 +6,7 @@ protocol MoviesLoading {
 
 struct MoviesLoader: MoviesLoading {
     // MARK: - NetworkClient
+    
     private let networkClient: NetworkRouting
     
     init(networkClient: NetworkRouting = NetworkClient()) {
@@ -13,6 +14,7 @@ struct MoviesLoader: MoviesLoading {
       }
     
     // MARK: - URL
+    
     private var mostPopularMoviesUrl: URL {
         guard let url = URL(string: "https://imdb-api.com/en/API/Top250Movies/k_zcuw1ytf") else {
             preconditionFailure("Unable to construct mostPopularMoviesUrl")
