+++ MovieQuizTests/MovieQuizViewControllerMock.swift
@@ -29,19 +29,9 @@ final class MovieQuizViewControllerMock: MovieQuizViewControllerProtocol {
     func showNetworkError(message: String) {
     
     }
-}
-
-final class MovieQuizPresenterTests: XCTestCase {
-    func testPresenterConvertModel() throws {
-        let viewControllerMock = MovieQuizViewControllerMock()
-        let sut = MovieQuizPresenter(viewController: viewControllerMock)
+    
+    func showAlert(model: AlertModel) {
         
-        let emptyData = Data()
-        let question = QuizQuestion(image: emptyData, text: "Question Text", correctAnswer: true)
-        let viewModel = sut.convert(model: question)
-        
-        XCTAssertNotNil(viewModel.image)
-        XCTAssertEqual(viewModel.question, "Question Text")
-        XCTAssertEqual(viewModel.questionNumber, "1/10")
     }
 }
+
